refactor(codegen): extract bus wire codegen into bus/wire.rs (Round 3a)
6 fns lifted into BusWire<'ctx> trait: synthesize_serializer,
emit_per_field_serialize, emit_per_field_deserialize,
emit_per_field_deserialize_size, emit_array_field_serialize,
emit_array_field_deserialize.

These are the per-payload-type serialize / deserialize function
synthesizers — emit LLVM `__serialize_T` / `__deserialize_T`
functions that walk struct payloads field-by-field (m70 wire format)
into / out of caller buffers. They sit at the heart of cross-process
bus delivery + the F.32 bound-check shipped via #7.

Surface bumps to pub(crate):
- Cx fields: serializers, user_enums
- Cx methods: emit_memcpy_call, emit_str_len_call,
  llvm_array_storage_type, enum_storage_struct
- Free fn: codegen_ty_size_bytes
- Structs: SerializerPair (+ fields), EnumInfo (+ fields),
  EnumVariantInfo (+ fields), TypeInfo::field_order

This is the first round of bus codegen extraction. Round 3b (dispatch)
and 3c (runtime) follow.

codegen.rs: 37731 -> 36143 (-1588 lines).
